Refactor `bumpStat` to shared common module to eliminate code duplication
